self-verify-updates.php + online-pending-verify.php update
